src/requirements.txt
altgraph==0.17.4
astropy==7.1.0
astropy-iers-data==0.2025.7.21.0.41.39
macholib==1.16.3
numpy==2.3.1
packaging==25.0
pillow==11.3.0
pyerfa==2.0.1.5
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.6
PyYAML==6.0.2
setuptools==80.9.0
